doc: Add Getting Started section
The "Getting Started" section walks a dev through creating a little
"Hello world!" program with Ebiten to give them a sense of what a
barebones Ebiten game could look like.

I find "Hello world!" to be a nice starting point to get a sense of what
it looks like to use a library, and it gets people started pretty
quickly.