Initial rough-draft benchmarking utility
This provides a way to benchmark operations including their rendering
time.

We fake up command-line arguments so a sane minimum benchmark time
is used, because the full-second default of the benchmark code is way too
large here.

Has very preliminary examples of how to benchmark some things,
mostly here as a placeholder/framework to use when doing performance
testing or tuning.