graphics: Bug fix: fullscreen rendering was wrong
After fad65f2f5d2166dd64b71686b7a4df4a032eb097, the viewport size
is always same as the framebuffer size. The 'screen' image size
was the actual rendering region size, and padding was not
considered in the size. The padding is added on the fullscreen
mode.

This fix adds the extra padding values to the 'screen' image size
so that the 'screen' framebuffer will be same size as the monitor
size.

Fixes #693