graphics: Use math.Ceil for the screen framebuffer size
There are glitches on some mobile devices like Xperia. As the
screen framebuffer size might not fit with the actual GL surface
size, we guessed that is the cause of the glitches.

By using math.Ceil, the screen framebuffer will be a little bigger
and can cover the GL surface.