graphics: Bug fix: Resolve pixels before the screen is cleared
Fixes #792

Clearing the screen is so special on some environments (like macOS
Metal) that other operations on other images can be invalidated.

This fixes the issue by resolving the pixels of the offscreen that
is used as a render source for the screen before the screen is
cleared. We are not sure this is a correct way, but this actually
fixes the problem on macOS.