internal/mipmap: reland: do not consider filter to determine mipmap level
This is a reland of 42833614fb6f24e5daa829df47f446fcd0f970ac.

Now the filter argument is not used (a4e9a05b14997f431457955ec7f5cf31e796e112),
the filter value can be the nearest filter even though the shader is
for the linear filter.

As long as `canSkipMipmap` is set correctly, we don't have to consider
the parameter `filter`. We should ignore it.

Updates #2364