add Image.SubImageInto method
Instead of allocating the result on heap for the user,
allow the caller side to decide whether it should be heap-allocated
or stack-allocated.

This new method allows us to have a zero-alloc SubImage routine
for the use cases where we don't want to make an extra allocation.

The performance difference is measurable (comparison is done with benchstat):

	name             old time/op    new time/op    delta
	SubImage-12     203ns ± 2%      33ns ± 2%   -83.91%  (p=0.000 n=20+19)

	name             old alloc/op   new alloc/op   delta
	SubImage-12      112B ± 0%        0B       -100.00%  (p=0.000 n=20+20)

Using SubImage is `~203ns`, SubImageInto is `~33ns`.
SubImage does 1 allocation, SubImageInto does none.

SubImage is rewritten in a way to avoid the code duplication.
It should work identically to the previous contract (it returns nil
if `i` is disposed, etc.)

Fixes #2902