internal/ui: bug fix: the window size callback was not called on macOS
This is basically a revert of e150745598f246f4a1236f6da0f87dcc2dad16b7.

As the window size callback was not called, ebiten.WindowSize didn't
work correctly for a while.

Even after this change, the issue #2257 could not be reproduced. This
is mysterious, but probably macOS itself has updated.

Updates #2257
Closes #3029