internal/graphicsdriver/metal: bug fix: fix possible lost-wake issue
As the fence value could be updated regardless of the mutex, the
fence value can be updated just befor the receiver reaches Wait and
just after confirming the condition false. This didn't cause a critical
issue like dead lock as the fence value would be updated eventually,
but this is not efficient.

Updates #3278