internal/graphicsdriver/metal: separate command buffers into two
For efficiency, this change separatens the current command buffer
into two: for non-screen and for screen. It's because a drawable will
not be released as long as the CommandBuffer referencing it is alive,
it is more efficient to separate CommandBuffers that use the drawable
from those that do not.

Updates #3278