cmd/ebitenmobile: save GPU resources at onPause and restore GPU resources on context lost
This change adds a logic to save GPU resources at GLSurfaceView.onPause
as a precaution, and restore GPU resources from it when a context lost
happens. This can fix issues where a context lost often occurs when an
interstitial ad is shown.

This change doesn't add any new APIs and do this resource reservation
automatically.

Closes #3215